extract fixture.
